Move activity_pool and server to the separate plugin folder
